[FIX] wallet schema 수정 (float -> double) + 불필요한 코드 제거

+++ src/main/java/com/konkuk/kubit/domain/Wallet.java
@@ -27,8 +27,8 @@ public class Wallet {
     @JoinColumn(name="marketCode")
     private Market marketCode;
 
-    private float quantity;
+    private double quantity;
 
-    private float totalPrice;
+    private double totalPrice;
 }
